recommender_api/requirements.txt
asyncio==3.4.3
blinker==1.6.2
blis>=0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
de-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.5.0/de_core_news_lg-3.5.0-py3-none-any.whl#sha256=fcce8e827230b14517e361ca59bf2bd3faaec2752007eb5b5b43b6093d1381f0
Flask==2.3.2
html2text==2020.1.16
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
numpy==1.25.0
packaging==23.1
pandas==2.0.2
pathy==0.10.1
preshed==3.0.8
psycopg2==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
threadpoolctl==3.1.0
tqdm==4.65.0
typer>=0.7.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wasabi==1.1.2
websockets==11.0.3
Werkzeug==2.3.6
